requirements.txt
mysql-connector-python >= 8.0.28

django-mysql>=3.10.0,<3.11.0

Django>=3.2,<4.0

djangorestframework>=3.12.2,<3.13.0

flake8>=3.8.4,<3.9.0

gunicorn>=20.0.4,<20.1.0

s3fs

coverage>=5.5,<6.0

django-cors-headers>=3.7.0,<3.8.0

ddt>=1.4.2,<1.5.0

django-model-utils>=4.1.1,<4.2.0

requests>=2.25.1,<2.26.0
